Move ComponentOutOfRangeError to the backend since it's API is specific
